+++ src/main/java/roomescape/domain/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure;
+package roomescape.domain;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -8,8 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationDate;
 
 @Repository
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
