+++ src/main/java/roomescape/domain/ReservationTimeRepository.java
@@ -1,12 +1,11 @@
-package roomescape.infrastructure;
+package roomescape.domain;
 
 import java.time.LocalTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationTime;
 
 @Repository
 public interface ReservationTimeRepository extends JpaRepository<ReservationTime, Long> {
-    
+
     boolean existsByStartAt(LocalTime parsedTime);
 }
