+++ src/main/java/roomescape/domain/ReservationWaitingRepository.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure;
+package roomescape.domain;
 
 import java.util.List;
 import org.springframework.data.domain.Page;
@@ -8,7 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationWaiting;
 
 @Repository
 public interface ReservationWaitingRepository extends JpaRepository<ReservationWaiting, Long> {
