+++ src/main/java/roomescape/domain/ThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure;
+package roomescape.domain;
 
 
 import java.time.LocalDate;
@@ -8,7 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Theme;
 
 @Repository
 public interface ThemeRepository extends JpaRepository<Theme, Long> {
