+++ src/main/java/roomescape/service/ThemeService.java
@@ -7,9 +7,9 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.domain.ReservationRepository;
 import roomescape.domain.Theme;
-import roomescape.infrastructure.ReservationRepository;
-import roomescape.infrastructure.ThemeRepository;
+import roomescape.domain.ThemeRepository;
 import roomescape.service.exception.ReservationExistsException;
 import roomescape.service.request.ThemeAppRequest;
 import roomescape.service.response.ThemeAppResponse;
