+++ src/test/java/roomescape/service/MemberAuthServiceTest.java
@@ -1,26 +1,31 @@
 package roomescape.service;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.when;
+import static roomescape.Fixture.VALID_USER_EMAIL;
+import static roomescape.Fixture.VALID_USER_NAME;
+import static roomescape.Fixture.VALID_USER_PASSWORD;
+
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.domain.*;
-import roomescape.infrastructure.MemberRepository;
+import roomescape.domain.Member;
+import roomescape.domain.MemberEmail;
+import roomescape.domain.MemberName;
+import roomescape.domain.MemberPassword;
+import roomescape.domain.MemberRepository;
+import roomescape.domain.MemberRole;
 import roomescape.service.request.MemberSignUpAppRequest;
-import roomescape.service.response.SignupMemberAppResponse;
 import roomescape.service.response.MemberAppResponse;
-
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
-import static roomescape.Fixture.*;
+import roomescape.service.response.SignupMemberAppResponse;
 
 @ExtendWith(MockitoExtension.class)
 class MemberAuthServiceTest {
