+++ src/test/java/roomescape/service/ReservationWaitingServiceTest.java
@@ -25,13 +25,13 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.domain.Member;
+import roomescape.domain.MemberRepository;
 import roomescape.domain.MemberRole;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationDate;
+import roomescape.domain.ReservationRepository;
 import roomescape.domain.ReservationWaiting;
-import roomescape.infrastructure.MemberRepository;
-import roomescape.infrastructure.ReservationRepository;
-import roomescape.infrastructure.ReservationWaitingRepository;
+import roomescape.domain.ReservationWaitingRepository;
 import roomescape.service.exception.PastReservationException;
 import roomescape.service.request.ReservationWaitingAppRequest;
 import roomescape.service.response.ReservationWaitingAppResponse;
