+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -1,5 +1,8 @@
 package roomescape.service;
 
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.Mockito.when;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -9,13 +12,10 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.infrastructure.ReservationRepository;
-import roomescape.infrastructure.ThemeRepository;
+import roomescape.domain.ReservationRepository;
+import roomescape.domain.ThemeRepository;
 import roomescape.service.exception.ReservationExistsException;
 import roomescape.service.request.ThemeAppRequest;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class ThemeServiceTest {
